Made global state to manage selected tabs
