+++ src/components/tabs/index.jsx
@@ -1,15 +1,11 @@
 import { useState } from 'react';
+import { useContext } from 'react';
+import { TabContext } from '../../context/TabContext';
 
-const navItems = [
-    { name: 'View all', href: '#' },
-    { name: 'Management', href: '#' },
-    { name: 'Design', href: '#' },
-    { name: 'Photographers', href: '#' },
-    { name: 'Operations', href: '#' }
-];
 
 export default function Tabs() {
-    const [activeItem, setActiveItem] = useState('View all');
+    // using the tab's global context 
+    const {navItems, activeItem, setActiveItem} = useContext(TabContext);
 
     return (
         <nav className="w-full overflow-x-auto flex justify-start sm:justify-center items-center px-2 sm:px-4 lg:px-8">
@@ -19,7 +15,7 @@ export default function Tabs() {
                         key={item.name}
                         href={item.href}
                         onClick={() => setActiveItem(item.name)}
-                        className={`font-inter border-[1.2px] py-2 px-4 text-sm sm:text-md lg:text-lg text-black font-bold transition-colors rounded-full ${
+                        className={`border-[1.2px] py-2 px-4 text-sm sm:text-md lg:text-lg text-black font-bold transition-colors rounded-full ${
                             activeItem === item.name
                                 ? 'border-black text-black'
                                 : 'border-transparent text-gray-600 hover:text-gray-900'
