+++ src/main.jsx
@@ -3,11 +3,14 @@ import { createRoot } from "react-dom/client"
 import "./index.css"
 import App from "./App.jsx"
 import { BrowserRouter } from "react-router"
+import { TabContextProvider } from "./context/TabContext.jsx"
 
 createRoot(document.getElementById("root")).render(
 	<StrictMode>
-		<BrowserRouter>
-			<App />
-		</BrowserRouter>
+		<TabContextProvider>
+			<BrowserRouter>
+				<App />
+			</BrowserRouter>
+		</TabContextProvider>
 	</StrictMode>
 )
