+++ src/pages/portfolio/index.jsx
@@ -7,7 +7,7 @@ const PortfolioPage = () => {
     <>
         {/* top most section  */}
         <div className="mt-11 flex flex-col justify-center items-center gap-y-4 px-4 sm:px-6 lg:px-8">
-          <span className="text-center font-inter font-bold text-3xl border-[1.2px] border-black rounded-full px-4 py-2 w-auto sm:text-4xl lg:text-5xl">
+          <span className="text-center font-bold text-3xl border-[1.2px] border-black rounded-full px-4 py-2 w-auto sm:text-4xl lg:text-5xl">
             Meet our team
           </span>
           <span className="text-md opacity-60 sm:text-lg lg:text-xl lg:max-w-2xl text-center">
